Arreglo de NewsDetail
Luego de ver el Epic Link de novedades entendi que la confusion fue
xq Tomi habia hecho la pagina de Detalles de novedades pero no le
correspondia, y yo trabaje sobre esa. Despues Manu hizo el llamado a la
API en la pagina de detalles, que me correspondia a mi. Lo arregle
modificando el archivo de manu y reorganizando las carpetas y para
mostrar la evidencia use mock data.
